+++ vck-openid/src/commonTest/kotlin/at/asitplus/wallet/lib/openid/PreRegisteredClientTest.kt
@@ -20,7 +20,9 @@ import io.kotest.matchers.string.shouldContain
 import io.kotest.matchers.string.shouldNotContain
 import io.kotest.matchers.string.shouldStartWith
 import io.kotest.matchers.types.shouldBeInstanceOf
-import io.ktor.http.*
+import io.ktor.http.URLBuilder
+import io.ktor.http.Url
+import kotlin.collections.set
 
 class PreRegisteredClientTest : FreeSpec({
 
@@ -83,7 +85,7 @@ class PreRegisteredClientTest : FreeSpec({
 
         val result = verifierOid4vp.validateAuthnResponse(authnResponse.url)
             .shouldBeInstanceOf<AuthnResponseResult.Success>()
-        result.vp.verifiableCredentials.shouldNotBeEmpty()
+        result.vp.timelyVerifiableCredentials.shouldNotBeEmpty()
 
         verifierOid4vp.createAuthnRequest(
             defaultRequestOptions, OpenId4VpVerifier.CreationOptions.Query(walletUrl)
@@ -114,7 +116,7 @@ class PreRegisteredClientTest : FreeSpec({
 
         val result = verifierOid4vp.validateAuthnResponse(authnResponse.url)
             .shouldBeInstanceOf<AuthnResponseResult.Success>()
-        result.vp.verifiableCredentials.shouldNotBeEmpty()
+        result.vp.timelyVerifiableCredentials.shouldNotBeEmpty()
         result.state.shouldBe(expectedState)
     }
 
@@ -204,7 +206,7 @@ class PreRegisteredClientTest : FreeSpec({
 
         val result = verifierOid4vp.validateAuthnResponse(authnResponse.params.formUrlEncode())
             .shouldBeInstanceOf<AuthnResponseResult.Success>()
-        result.vp.verifiableCredentials.shouldNotBeEmpty()
+        result.vp.timelyVerifiableCredentials.shouldNotBeEmpty()
     }
 
     "test with direct_post_jwt" {
@@ -229,7 +231,7 @@ class PreRegisteredClientTest : FreeSpec({
 
         val result = verifierOid4vp.validateAuthnResponse(authnResponse.params.formUrlEncode())
             .shouldBeInstanceOf<AuthnResponseResult.Success>()
-        result.vp.verifiableCredentials.shouldNotBeEmpty()
+        result.vp.timelyVerifiableCredentials.shouldNotBeEmpty()
     }
 
     "test with deserializing" {
@@ -248,7 +250,7 @@ class PreRegisteredClientTest : FreeSpec({
 
         val result = verifierOid4vp.validateAuthnResponse(authnResponseParams)
             .shouldBeInstanceOf<AuthnResponseResult.Success>()
-        result.vp.verifiableCredentials.shouldNotBeEmpty()
+        result.vp.timelyVerifiableCredentials.shouldNotBeEmpty()
     }
 
     "test specific credential" {
@@ -262,8 +264,8 @@ class PreRegisteredClientTest : FreeSpec({
 
         val result = verifierOid4vp.validateAuthnResponse(authnResponse.url)
             .shouldBeInstanceOf<AuthnResponseResult.Success>()
-        result.vp.verifiableCredentials.shouldNotBeEmpty()
-        result.vp.verifiableCredentials.forEach {
+        result.vp.timelyVerifiableCredentials.shouldNotBeEmpty()
+        result.vp.timelyVerifiableCredentials.map { it.vcJws }.forEach {
             it.vc.credentialSubject.shouldBeInstanceOf<AtomicAttribute2023>()
         }
     }
@@ -278,8 +280,8 @@ class PreRegisteredClientTest : FreeSpec({
 
         val result = verifierOid4vp.validateAuthnResponse(authnResponse.url)
             .shouldBeInstanceOf<AuthnResponseResult.Success>()
-        result.vp.verifiableCredentials.shouldNotBeEmpty()
-        result.vp.verifiableCredentials.forEach {
+        result.vp.timelyVerifiableCredentials.shouldNotBeEmpty()
+        result.vp.timelyVerifiableCredentials.map { it.vcJws }.forEach {
             it.vc.credentialSubject.shouldBeInstanceOf<AtomicAttribute2023>()
         }
     }
@@ -310,8 +312,8 @@ class PreRegisteredClientTest : FreeSpec({
 
         val result = verifierOid4vp.validateAuthnResponse(authnResponse.url)
             .shouldBeInstanceOf<AuthnResponseResult.Success>()
-        result.vp.verifiableCredentials.shouldNotBeEmpty()
-        result.vp.verifiableCredentials.forEach {
+        result.vp.timelyVerifiableCredentials.shouldNotBeEmpty()
+        result.vp.timelyVerifiableCredentials.map { it.vcJws }.forEach {
             it.vc.credentialSubject.shouldBeInstanceOf<AtomicAttribute2023>()
         }
     }
@@ -336,8 +338,8 @@ class PreRegisteredClientTest : FreeSpec({
 
         val result = verifierOid4vp.validateAuthnResponse(authnResponse.url)
             .shouldBeInstanceOf<AuthnResponseResult.Success>()
-        result.vp.verifiableCredentials.shouldNotBeEmpty()
-        result.vp.verifiableCredentials.forEach {
+        result.vp.timelyVerifiableCredentials.shouldNotBeEmpty()
+        result.vp.timelyVerifiableCredentials.map { it.vcJws }.forEach {
             it.vc.credentialSubject.shouldBeInstanceOf<AtomicAttribute2023>()
         }
     }