+++ vck/src/commonMain/kotlin/at/asitplus/wallet/lib/agent/validation/CredentialTimelinessValidator.kt
@@ -4,6 +4,9 @@ import at.asitplus.wallet.lib.agent.SubjectCredentialStore
 import at.asitplus.wallet.lib.agent.validation.mdoc.MdocTimelinessValidator
 import at.asitplus.wallet.lib.agent.validation.sdJwt.SdJwtTimelinessValidator
 import at.asitplus.wallet.lib.agent.validation.vcJws.VcJwsTimelinessValidator
+import at.asitplus.wallet.lib.data.VerifiableCredentialJws
+import at.asitplus.wallet.lib.data.VerifiableCredentialSdJwt
+import at.asitplus.wallet.lib.iso.IssuerSigned
 import kotlinx.datetime.Clock
 import kotlin.time.Duration
 import kotlin.time.Duration.Companion.seconds
@@ -24,29 +27,32 @@ data class CredentialTimelinessValidator(
         clock = clock
     ),
 ) {
+    operator fun invoke(issuerSigned: IssuerSigned) = CredentialTimelinessValidationSummary.Mdoc(
+        mdocTimelinessValidator(issuerSigned),
+    )
+
+    operator fun invoke(sdJwt: VerifiableCredentialSdJwt) = CredentialTimelinessValidationSummary.SdJwt(
+        sdJwtTimelinessValidator(sdJwt),
+    )
+
+    operator fun invoke(vcJws: VerifiableCredentialJws) = CredentialTimelinessValidationSummary.VcJws(
+        vcJwsTimelinessValidator(vcJws = vcJws),
+    )
+
     operator fun invoke(storeEntry: SubjectCredentialStore.StoreEntry) = when (storeEntry) {
-        is SubjectCredentialStore.StoreEntry.Iso -> invoke(CredentialWrapper.Mdoc(storeEntry.issuerSigned))
-        is SubjectCredentialStore.StoreEntry.SdJwt -> invoke(CredentialWrapper.SdJwt(storeEntry.sdJwt))
-        is SubjectCredentialStore.StoreEntry.Vc -> invoke(CredentialWrapper.VcJws(storeEntry.vc))
+        is SubjectCredentialStore.StoreEntry.Iso -> invoke(storeEntry)
+        is SubjectCredentialStore.StoreEntry.SdJwt -> invoke(storeEntry)
+        is SubjectCredentialStore.StoreEntry.Vc -> invoke(storeEntry)
     }
+    operator fun invoke(storeEntry: SubjectCredentialStore.StoreEntry.Vc) = invoke(storeEntry.vc)
+    operator fun invoke(storeEntry: SubjectCredentialStore.StoreEntry.SdJwt) = invoke(storeEntry.sdJwt)
+    operator fun invoke(storeEntry: SubjectCredentialStore.StoreEntry.Iso) = invoke(storeEntry.issuerSigned)
 
-    operator fun invoke(credentialWrapper: CredentialWrapper): CredentialTimelinessValidationSummary {
-        return when (credentialWrapper) {
-            is CredentialWrapper.Mdoc -> CredentialTimelinessValidationSummary.Mdoc(
-                credential = credentialWrapper,
-                mdocTimelinessValidator.invoke(credentialWrapper.issuerSigned),
-            )
 
-            is CredentialWrapper.SdJwt -> CredentialTimelinessValidationSummary.SdJwt(
-                credential = credentialWrapper,
-                sdJwtTimelinessValidator(credentialWrapper.sdJwt),
-            )
-
-            is CredentialWrapper.VcJws -> CredentialTimelinessValidationSummary.VcJws(
-                credential = credentialWrapper,
-                vcJwsTimelinessValidator(vcJws = credentialWrapper.verifiableCredentialJws),
-            )
-        }
+    operator fun invoke(credentialWrapper: CredentialWrapper): CredentialTimelinessValidationSummary = when (credentialWrapper) {
+        is CredentialWrapper.Mdoc -> invoke(credentialWrapper.issuerSigned)
+        is CredentialWrapper.SdJwt -> invoke(credentialWrapper.sdJwt)
+        is CredentialWrapper.VcJws -> invoke(credentialWrapper.verifiableCredentialJws)
     }
 }