+++ vck/src/commonMain/kotlin/at/asitplus/wallet/lib/agent/validation/sdJwt/SdJwtTimelinessValidationDetails.kt
@@ -4,7 +4,7 @@ import at.asitplus.wallet.lib.agent.validation.common.EntityExpiredError
 import at.asitplus.wallet.lib.agent.validation.common.EntityNotYetValidError
 import kotlinx.datetime.Instant
 
-data class SdJwtTimelinessValidationSummary(
+data class SdJwtTimelinessValidationDetails(
     val evaluationTime: Instant,
     val jwsExpiredError: EntityExpiredError?,
     val jwsNotYetValidError: EntityNotYetValidError?,