+++ vck/src/commonMain/kotlin/at/asitplus/wallet/lib/data/IsoDocumentParsed.kt
@@ -1,7 +1,7 @@
 package at.asitplus.wallet.lib.data
 
 import at.asitplus.KmmResult
-import at.asitplus.wallet.lib.agent.validation.mdoc.MdocTimelinessValidationSummary
+import at.asitplus.wallet.lib.agent.validation.CredentialTimelinessValidationSummary
 import at.asitplus.wallet.lib.data.rfc.tokenStatusList.primitives.TokenStatus
 import at.asitplus.wallet.lib.iso.IssuerSignedItem
 import at.asitplus.wallet.lib.iso.MobileSecurityObject
@@ -15,5 +15,5 @@ data class IsoDocumentParsed(
     val validItems: List<IssuerSignedItem> = listOf(),
     val invalidItems: List<IssuerSignedItem> = listOf(),
     val tokenStatus: KmmResult<TokenStatus>?,
-    val timelinessValidationSummary: MdocTimelinessValidationSummary,
+    val timelinessValidationSummary: CredentialTimelinessValidationSummary.Mdoc,
 )